refactor(renderer): factorise environment columns
