--- packages/renderer/src/lib/container/ContainerColumnEnvironment.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-import ProviderInfo from '../ui/ProviderInfo.svelte';
-import type { ContainerGroupInfoUI, ContainerInfoUI } from './ContainerInfoUI';
-
-export let object: ContainerInfoUI | ContainerGroupInfoUI;
-</script>
-
-<ProviderInfo provider={object.engineName} context={object.engineId} />
