--- packages/renderer/src/lib/image/ImageColumnEnvironment.spec.ts
@@ -1,52 +1,0 @@
-/**********************************************************************
- * Copyright (C) 2023-2024 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ***********************************************************************/
-
-import '@testing-library/jest-dom/vitest';
-
-import { render, screen } from '@testing-library/svelte';
-import { expect, test } from 'vitest';
-
-import ImageIcon from '../images/ImageIcon.svelte';
-import ImageColumnEnvironment from './ImageColumnEnvironment.svelte';
-import type { ImageInfoUI } from './ImageInfoUI';
-
-test('Expect simple column styling', async () => {
-  const image: ImageInfoUI = {
-    id: 'my-image',
-    shortId: '',
-    name: '',
-    engineId: '',
-    engineName: 'podman',
-    tag: '',
-    createdAt: 0,
-    age: '',
-    arch: '',
-    size: 0,
-    humanSize: '',
-    base64RepoTag: '',
-    selected: false,
-    status: 'UNUSED',
-    icon: ImageIcon,
-    badges: [],
-    digest: 'sha256:1234567890',
-  };
-  render(ImageColumnEnvironment, { object: image });
-
-  const text = screen.getByText(image.engineName);
-  expect(text).toBeInTheDocument();
-});
