--- packages/renderer/src/lib/image/ImageColumnEnvironment.svelte
@@ -1,12 +1,0 @@
-<script lang="ts">
-import ProviderInfo from '../ui/ProviderInfo.svelte';
-import type { ImageInfoUI } from './ImageInfoUI';
-
-interface Props {
-  object: ImageInfoUI;
-}
-
-let { object }: Props = $props();
-</script>
-
-<ProviderInfo provider={object.engineName} context={object.engineId} />
